refactor: rename parachain to template to better describe intention
